requirements.txt
langchain
sentence-transformers
faiss-cpu
transformers
python-dotenv
pickle5
